[clang-tidy] Rename 'google-build-namespaces' check to 'misc-anonymous-namespace-in-header'
